test: split unittest
